+++ test/components/datepicker/datepicker.e2e-spec.js
@@ -486,18 +486,17 @@ describe('Datepicker Destroy Mask Tests', () => {
 
   it('Should still mask after destroy', async () => {
     await browser.driver.sleep(config.sleepShort);
-    const inputEl = await element(by.id('dp1'));
-    inputEl.clear();
-    inputEl.sendKeys('101020011221AM');
+    await element(by.id('dp1')).clear();
+    await element(by.id('dp1')).sendKeys('101020011221AM');
 
-    expect(inputEl.getAttribute('value')).toEqual('10/10/2001 12:21 AM');
-    inputEl.clear();
+    expect(await element(by.id('dp1')).getAttribute('value')).toEqual('10/10/2001 12:21 AM');
+    await element(by.id('dp1')).clear();
 
     await element(by.id('btn-update')).click();
     await browser.driver.sleep(config.sleepShort);
-    inputEl.sendKeys('101020011221AM');
+    await element(by.id('dp1')).sendKeys('101020011221AM');
 
-    expect(inputEl.getAttribute('value')).toEqual('10/10/2001 12:21 AM');
+    expect(await element(by.id('dp1')).getAttribute('value')).toEqual('10/10/2001 12:21 AM');
   });
 });
 
@@ -1227,5 +1226,56 @@ describe('Datepicker Time in Cs-Cz Format Tests', () => {
     expect(await element(by.id('dp1')).getAttribute('value')).toEqual('26.02.2016 9:15 PM');
     expect(await element(by.id('dp2')).getAttribute('value')).toEqual('26.02.2016 14:15');
     expect(await element(by.id('dp3')).getAttribute('value')).toEqual('05.04.2018 16:15');
+  });
+});
+
+describe('Datepicker specific locale/langauge tests', () => {
+  beforeEach(async () => {
+    await utils.setPage('/components/datepicker/test-two-locales-same-page');
+  });
+
+  it('Should be able to use current locale', async () => {
+    const datepickerEl = await element(by.id('date-field-normal'));
+    await element(by.css('#date-field-normal + .icon')).click();
+
+    expect(await element(by.css('.hyperlink.today')).getText()).toEqual('Today');
+    await element(by.css('.hyperlink.today')).click();
+
+    const testDate = new Date();
+    testDate.setHours(0);
+    testDate.setMinutes(0);
+    testDate.setSeconds(0);
+
+    expect(await datepickerEl.getAttribute('value')).toEqual(testDate.toLocaleDateString('en-US'));
+  });
+
+  it('Should be able to use non current locale', async () => {
+    const datepickerEl = await element(by.id('date-field-danish'));
+    await element(by.css('#date-field-danish + .icon')).click();
+
+    expect(await element(by.css('.hyperlink.today')).getText()).toEqual('I dag');
+    await element(by.css('.hyperlink.today')).click();
+
+    const testDate = new Date();
+    testDate.setHours(0);
+    testDate.setMinutes(0);
+    testDate.setSeconds(0);
+
+    expect(await datepickerEl.getAttribute('value')).toEqual(`${testDate.getDate()}-${testDate.getMonth() + 1}-${testDate.getFullYear()}`);
+  });
+
+  it('Should be Able to use non current locale and a different langauge', async () => {
+    const datepickerEl = await element(by.id('date-field-sv-de'));
+    await element(by.css('#date-field-sv-de + .icon')).click();
+
+    expect(await element(by.css('.hyperlink.today')).getText()).toEqual('Idag');
+    await element(by.css('.hyperlink.today')).click();
+
+    const testDate = new Date();
+    testDate.setHours(0);
+    testDate.setMinutes(0);
+    testDate.setSeconds(0);
+
+    expect(await datepickerEl.getAttribute('value')).toEqual(`${testDate.getDate()}.${testDate.getMonth() + 1}.${testDate.getFullYear()}`);
   });
 });